+++ index.js
@@ -47,14 +47,14 @@ Negotiator.prototype.charsets = function charsets(available) {
   return preferredCharsets(this.request.headers['accept-charset'], available);
 };
 
-Negotiator.prototype.encoding = function encoding(available) {
-  var set = this.encodings(available);
+Negotiator.prototype.encoding = function encoding(available, options) {
+  var set = this.encodings(available, options);
   return set && set[0];
 };
 
-Negotiator.prototype.encodings = function encodings(available) {
+Negotiator.prototype.encodings = function encodings(available, options) {
   var preferredEncodings = loadModule('encoding').preferredEncodings;
-  return preferredEncodings(this.request.headers['accept-encoding'], available);
+  return preferredEncodings(this.request.headers['accept-encoding'], available, options);
 };
 
 Negotiator.prototype.language = function language(available) {