+++ test/encoding.js
@@ -2,6 +2,10 @@
 var assert = require('assert')
 var Negotiator = require('..')
 
+var clientPreference = { sortPreference: 'client'}
+var serverPreference = { sortPreference: 'server'}
+var clientThenServerPreference = { sortPreference: 'clientThenServer'}
+
 describe('negotiator.encoding()', function () {
   whenAcceptEncoding(undefined, function () {
     it('should return identity', function () {
@@ -112,6 +116,12 @@ describe('negotiator.encoding(array)', function () {
       assert.strictEqual(this.negotiator.encoding(['gzip']), 'gzip')
       assert.strictEqual(this.negotiator.encoding(['compress', 'gzip']), 'gzip')
     })
+
+    it('clientThenServerPreference: should return server-preferred encoding', function () {
+      assert.strictEqual(this.negotiator.encoding(['identity'], clientThenServerPreference), 'identity')
+      assert.strictEqual(this.negotiator.encoding(['gzip'], clientThenServerPreference), 'gzip')
+      assert.strictEqual(this.negotiator.encoding(['compress', 'gzip'], clientThenServerPreference), 'compress')
+    })
   })
 
   whenAcceptEncoding('*, gzip;q=0', function () {
@@ -204,6 +214,23 @@ describe('negotiator.encoding(array)', function () {
     it('should return most client-preferred encoding', function () {
       assert.strictEqual(this.negotiator.encoding(['gzip']), 'gzip')
       assert.strictEqual(this.negotiator.encoding(['compress', 'identity']), 'identity')
+    })
+  })
+
+  whenAcceptEncoding('gzip;q=0.9, sdhc, br;q=0.9', function () {
+    it('should return best server-preferred encoding of equal client-preferred encodings', function () {
+      assert.strictEqual(this.negotiator.encoding(['compress', 'br', 'gzip']), 'gzip')
+      assert.strictEqual(this.negotiator.encoding(['compress', 'gzip', 'br']), 'gzip')
+    })
+
+    it('should return best server-preferred encoding of equal client-preferred encodings', function () {
+      assert.strictEqual(this.negotiator.encoding(['compress', 'br', 'gzip'], clientThenServerPreference), 'br')
+      assert.strictEqual(this.negotiator.encoding(['compress', 'gzip', 'br'], clientThenServerPreference), 'gzip')
+    })
+
+    it('should return best server-preferred encoding of equal client-preferred encodings', function () {
+      assert.strictEqual(this.negotiator.encoding(['compress', 'br', 'gzip'], serverPreference), 'br')
+      assert.strictEqual(this.negotiator.encoding(['compress', 'gzip', 'br'], serverPreference), 'gzip')
     })
   })
 })
@@ -316,7 +343,13 @@ describe('negotiator.encodings(array)', function () {
     it('should prefer gzip', function () {
       assert.deepEqual(this.negotiator.encodings(['identity']), ['identity'])
       assert.deepEqual(this.negotiator.encodings(['gzip']), ['gzip'])
-      assert.deepEqual(this.negotiator.encodings(['compress', 'gzip']), ['gzip', 'compress'])
+      assert.deepEqual(this.negotiator.encodings(['compress', 'deflate', 'gzip']), ['gzip', 'compress', 'deflate'])
+    })
+
+    it('clientThenServerPreference: should return server preferred encodings', function () {
+      assert.deepEqual(this.negotiator.encodings(['identity'], clientThenServerPreference), ['identity'])
+      assert.deepEqual(this.negotiator.encodings(['gzip'], clientThenServerPreference), ['gzip'])
+      assert.deepEqual(this.negotiator.encodings(['compress', 'deflate', 'gzip'], clientThenServerPreference), ['compress', 'deflate', 'gzip'])
     })
   })
 
@@ -403,12 +436,98 @@ describe('negotiator.encodings(array)', function () {
       assert.deepEqual(this.negotiator.encodings(['deflate']), ['deflate'])
       assert.deepEqual(this.negotiator.encodings(['deflate', 'gzip']), ['deflate', 'gzip'])
     })
+
+    it('serverPreference: should ignore client quality levels', function () {
+      assert.deepEqual(this.negotiator.encodings(['deflate', 'gzip'], serverPreference), ['deflate', 'gzip'])
+      assert.deepEqual(this.negotiator.encodings(['gzip', 'deflate'], serverPreference), ['gzip', 'deflate'])
+    })
   })
 
   whenAcceptEncoding('gzip;q=0.8, identity;q=0.5, *;q=0.3', function () {
     it('should return client-preferred encodings', function () {
       assert.deepEqual(this.negotiator.encodings(['gzip']), ['gzip'])
       assert.deepEqual(this.negotiator.encodings(['identity', 'gzip', 'compress']), ['gzip', 'identity', 'compress'])
+    })
+  })
+
+  whenAcceptEncoding('not;q=0, med1;q=0.9, med2;q=0.9, high, *;q=0.8', function () {
+    it('clientPreference', function () {
+      assert.deepEqual(this.negotiator.encodings(['high', 'not', 'med2', 'other', 'med1'], clientPreference), ['high', 'med1', 'med2', 'other'])
+      assert.deepEqual(this.negotiator.encodings(['high', 'not', 'med1', 'other', 'med2'], clientPreference), ['high', 'med1', 'med2', 'other'])
+    })
+
+    it('clientThenServerPreference', function () {
+      assert.deepEqual(this.negotiator.encodings(['high', 'not', 'med2', 'other', 'med1'], clientThenServerPreference), ['high', 'med2', 'med1', 'other'])
+      assert.deepEqual(this.negotiator.encodings(['high', 'not', 'med1', 'other', 'med2'], clientThenServerPreference), ['high', 'med1', 'med2', 'other'])
+    })
+
+    it('serverPreference', function () {
+      assert.deepEqual(this.negotiator.encodings(['high', 'not', 'med2', 'other', 'med1'], serverPreference), ['high', 'med2', 'other', 'med1'])
+      assert.deepEqual(this.negotiator.encodings(['high', 'not', 'med1', 'other', 'med2'], serverPreference), ['high', 'med1', 'other', 'med2'])
+    })
+  })
+
+  whenAcceptEncoding('not;q=0, med1;q=0.9, med2;q=0.9, high, identity;q=0.9', function () {
+    it('clientPreference', function () {
+      assert.deepEqual(this.negotiator.encodings(['high', 'not', 'med2', 'other', 'med1', 'identity'], clientPreference), ['high', 'med1', 'med2', 'identity'])
+      assert.deepEqual(this.negotiator.encodings(['high', 'not', 'med1', 'other', 'med2', 'identity'], clientPreference), ['high', 'med1', 'med2', 'identity'])
+    })
+
+    it('clientThenServerPreference', function () {
+      assert.deepEqual(this.negotiator.encodings(['high', 'not', 'med2', 'other', 'med1', 'identity'], clientThenServerPreference), ['high', 'med2', 'med1', 'identity'])
+      assert.deepEqual(this.negotiator.encodings(['high', 'not', 'med1', 'other', 'med2', 'identity'], clientThenServerPreference), ['high', 'med1', 'med2', 'identity'])
+    })
+
+    it('serverPreference', function () {
+      assert.deepEqual(this.negotiator.encodings(['high', 'not', 'med2', 'other', 'med1', 'identity'], serverPreference), ['high', 'med2', 'med1', 'identity'])
+      assert.deepEqual(this.negotiator.encodings(['high', 'not', 'med1', 'other', 'med2', 'identity'], serverPreference), ['high', 'med1', 'med2', 'identity'])
+    })
+  })
+
+  whenAcceptEncoding('c;q=0.9, b;q=0.89, a;q=0.9, d;q=0.91, e;q=0.9', function () {
+    it('clientPreference', function () {
+      assert.deepEqual(this.negotiator.encodings(['b','c','d','e'], clientPreference), ['d', 'c', 'e', 'b']);
+      assert.deepEqual(this.negotiator.encodings(['c','d','e','b'], clientPreference), ['d', 'c', 'e', 'b']);
+      assert.deepEqual(this.negotiator.encodings(['d','e','b','c'], clientPreference), ['d', 'c', 'e', 'b']);
+      assert.deepEqual(this.negotiator.encodings(['e','b','c','d'], clientPreference), ['d', 'c', 'e', 'b']);
+
+      assert.deepEqual(this.negotiator.encodings(['a','c','e'], clientPreference), ['c', 'a', 'e']);
+      assert.deepEqual(this.negotiator.encodings(['c','e','a'], clientPreference), ['c', 'a', 'e']);
+      assert.deepEqual(this.negotiator.encodings(['e','a','c'], clientPreference), ['c', 'a', 'e']);
+
+      assert.deepEqual(this.negotiator.encodings(['identity','e','f'], clientPreference), ['e','identity']);
+      assert.deepEqual(this.negotiator.encodings(['f','identity','e'], clientPreference), ['e','identity']);
+      assert.deepEqual(this.negotiator.encodings(['e','f','identity'], clientPreference), ['e','identity']);
+    })
+
+    it('clientThenServerPreference', function () {
+      assert.deepEqual(this.negotiator.encodings(['b','c','d','e'], clientThenServerPreference), ['d', 'c', 'e', 'b']);
+      assert.deepEqual(this.negotiator.encodings(['c','d','e','b'], clientThenServerPreference), ['d', 'c', 'e', 'b']);
+      assert.deepEqual(this.negotiator.encodings(['d','e','b','c'], clientThenServerPreference), ['d', 'e', 'c', 'b']);
+      assert.deepEqual(this.negotiator.encodings(['e','b','c','d'], clientThenServerPreference), ['d', 'e', 'c', 'b']);
+
+      assert.deepEqual(this.negotiator.encodings(['a','c','e'], clientThenServerPreference), ['a', 'c', 'e']);
+      assert.deepEqual(this.negotiator.encodings(['c','e','a'], clientThenServerPreference), ['c', 'e', 'a']);
+      assert.deepEqual(this.negotiator.encodings(['e','a','c'], clientThenServerPreference), ['e', 'a', 'c']);
+
+      assert.deepEqual(this.negotiator.encodings(['identity','e','f'], clientThenServerPreference), ['e','identity']);
+      assert.deepEqual(this.negotiator.encodings(['f','identity','e'], clientThenServerPreference), ['e','identity']);
+      assert.deepEqual(this.negotiator.encodings(['e','f','identity'], clientThenServerPreference), ['e','identity']);
+    })
+
+    it('serverPreference', function () {
+      assert.deepEqual(this.negotiator.encodings(['b','c','d','e'], serverPreference), ['b','c','d','e']);
+      assert.deepEqual(this.negotiator.encodings(['c','d','e','b'], serverPreference), ['c','d','e','b']);
+      assert.deepEqual(this.negotiator.encodings(['d','e','b','c'], serverPreference), ['d','e','b','c']);
+      assert.deepEqual(this.negotiator.encodings(['e','b','c','d'], serverPreference), ['e','b','c','d']);
+
+      assert.deepEqual(this.negotiator.encodings(['a','c','e'], serverPreference), ['a', 'c', 'e']);
+      assert.deepEqual(this.negotiator.encodings(['c','e','a'], serverPreference), ['c', 'e', 'a']);
+      assert.deepEqual(this.negotiator.encodings(['e','a','c'], serverPreference), ['e', 'a', 'c']);
+
+      assert.deepEqual(this.negotiator.encodings(['identity','e','f'], serverPreference), ['identity', 'e']);
+      assert.deepEqual(this.negotiator.encodings(['f','identity','e'], serverPreference), ['identity', 'e']);
+      assert.deepEqual(this.negotiator.encodings(['e','f','identity'], serverPreference), ['e','identity']);
     })
   })
 })